Initial work on rewrite
